src/test/java/study/heap/p42628/SolutionTest.java
package study.heap.p42628;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;
import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

class SolutionTest {

    @Test
    void test1() {
        assertThat(new Solution().solution(new String[]{"I 16", "I -5643", "D -1", "D 1", "D 1", "I 123", "D -1"}))
                .isEqualTo(new int[]{0, 0});
    }

    @Test
    void test2() {
        assertThat(new Solution().solution(new String[]{"I -45", "I 653", "D 1", "I -642", "I 45", "I 97", "D 1", "D -1", "I 333"}))
                .isEqualTo(new int[]{333, -45});
    }
}